Remove branded large rocks and missiles of returning from a shop vault
Large rocks aren't usable for most species, as well as being even less
appropriate for giving brands to than the other ammunition. The returning
brand description is specific to throwing, and it doesn't seem appropriate
to change it for one rare vault's occasional use of returning bolts,
arrows and sling bullets.