Clean up a vault header by giving the vault a sane monster set.
With the introduction of Depths, people have been very surprised to
learn that there is a non-standard monster set for this vault given its
general indistinguishability from the native monster population, and are
even more surprised when they attempt to read what the vault is doing.

In the process, I spotted some now-defunct SUBST:s that have gone away.

I will be surprised if more than one or two people notice this change.