Some old Lair vaults/minivaults from carwin (#3819).
With a lot of modernisation, including some adjustment to the monster
sets.

These were originally submitted as a serial vault, but I feel that
placing them individually will probably help them feel more varied -
you're going to get a lot of "why are there so many of this monster on
this level" with this kind of thing as a serial vault.

Also, our lair layouts are a lot better than when this was first
submitted.