Don't waste time casting Animate Dead, usually.
If there's a visible corpse nearby, the behaviour is the same. If
there are no items, 'z' should not waste time/MP, while you can
still force a cast with 'Z', so that you can animate corpses that
are below other items and you can't see with ^X.

Sadly, this doesn't really work for Yred since there's no way to
force a use of an ability the way there is with a spell.