Really fix Gozag gold detection (#8584).
It was moving the wrong item to the top of the cell - I wanted
item_def::index(), not item_def::link.

Also, if the gold was the second item on the cell, it would get missed, and
if more than one identical item was being generated on the same tile before
and after the gold, it could also get missed.