Change Gr +1 melee apts to +0
Uneven melee aptitudes reduce decision space for choosing a weapon.
Since Gr are already fairly powerful, their high aptitudes were nerfed
instead of their lower aptitudes buffed. The original suggestion was for
the +1 aptitudes to be moved to -1, but wheals objected for flavor
reasons. The aptitudes can always be moved to -1 later.