Remove player burden and carrying capacity
Item inventory weights (based on item mass) generally don't lead to
meaningful decisions that justify the inventory juggling and interface
problems that come from having burden states. The 52-slot limit is a
better system for limiting inventory and providing inventory-related
decisions because it's not so fine-grained and doesn't require the
player to examine weights for each slot.  Work is ongoing to improve
the slot system by consolidating food types and handling strategic
consumables in a different way.