Remove SIZE_HUGE.
Do you know whether it was bigger than SIZE_GIANT or not? Chances
are no (it's bigger, to be clear). There are some small changes to
in constriction, though there aren't any constrictors at either size,
and knockback -- in particular, dragons will be less effective at
knocking back giants.

The other change is that Dragon Form will have slightly better EV,
as will monster dragons (but only while constricted or netted). The
change seems like it is rather minor; I tested it with 15 Dodging and
only had a difference of 3 EV.