Rework Confusing Touch
In its previous state, it was fiddly, misleading, and
kind of overpowered. So:

Raise it to level 2 (& remove it from Cantrips, replacing
it with Corona). Remove the stacking-duration effect &
increase the initial duration. Set damage of all hits to
0, and make the first successful (confusing) hit end the
effect.

This should still be useful, but in a level 2 kind of way.
We'll see how that pans out.