New mutation: Forlorn
This mutation acts as an anti-faith, making piety gain slower. This
probably doesn't work well with wretched stars, but it seems like a
good, long term, strategic effect.

This idea has been floated around before as a species mutation. It could
be added to one of the blander species, or given to a new species.