Remove Elephant Slugs (actually turn them to dummy monsters)
Basically this has all the effects of a removal, but the elephant slug
"stat block" remains as a dummy so Gastronok has a base type. The slugs
even get swapped out on load for ghosts.

One thing I've considered is turning Elephant Slugs into some other
monsters, such as worms. "Slug" conveys the fact it is slow better than
"Worm".