Don't purposely avoid acquiring things the player is carrying
So for example, if you had a lamp of fire in your inventory and acquired
misc, this code meant acquirement would purposely avoid giving you
another lamp of fire. If it was in your stash, however, acquirement
didn't care. This is obviously wrong since things like miscellaneous
acquirement assume you can get multiple of a single sub_type, and
violates the "don't make players do silly things before reading
?acquire" rule.

This might alter the standings of staff, misc, and jewellry acquirement,
but none of those are all stars likely to be totally broken by this
change.

I've traced this back to 1d0f57cbceb, with only a couple changes in
between, so this either traces to before stone soup or shortly after.