Add an option for more consistent prompt behaviour when changing jewellery
When enabled, you will always be prompted to choose a hand/tentacle when
equipping a ring, even if you currently have some empty slots, or if one ring
is cursed, etc. As a result, "Px>" should always result in "put ring x on
right hand", and never "put on ring x, then go downstairs", for example.