Show visible changes to status of stack toppers.
If two corpses were sitting on a visible square, ctrl-x would never
show the rotting of the visible top corpse; and the stash tracker
itself would show it as "rotten by now" even though we can actually
see the corpse rotting.  This also affects recharging rods since
the previous commit.

Now if the seen item and the remembered top item are "the same"
but not identical (because of rotting, recharging, etc.), we update
the stash anyway.