Rework the web effect of sack of spiders
Surrounding the player with webs requires the player to have a
controlled form of blink to use the sack safely in many
instances. Instead of placing nets everywhere within the radius, the
item now only attempts to place them on monsters and then triggers the
web. The chance that the monster will spend a turn trapped in a web
and vulnerable is smaller now, since the web triggers on the turn the
sack is used, instead of when the monster moves onto the web. This
change doesn't seem to be a significant buff to the item overall, so
leaving e.g. chance of web placement and number of uses unchanged for
now.