Add several people to CREDITS.txt
Generally using the version of the nam e in the commit metadata,
resolved in favour of the real name where that information was provided.

Let a dev know if you want us to list you differently, or if one of
your patches or tiles is included in Crawl but you aren't listed.