Give mercs better items.
They automatically get their items set as MAKE_GOOD_ITEM, which
might not be enough to make up for the mediocre weapon types they
tend to start with, but it's the simplest solution for now. They're
also marked as MF_HARD_RESET so that you can't not pay and then take
their stuff.

Also, fix compilation: I'll fix this (so that git bisect works) while
merging into master.