Don't check _accept_mutation in _get_mut_with_use.
Most of the code that was replaced by _get_mut_with_use did not care
whether the mutation was acceptible, and some of the code was broken by
the change.  In particular, _delete_random_slime_mutation indexed
you.mutations directly with the result, causing a crash when the
selected mutation was unacceptible [1]---and even if the crash were
fixed, the behaviour would be wrong because it would never select a
maxed-out slime mutation.

Move the _accept_mutation call out of _get_mut_with_use and into the one
caller that needed it, _get_random_mutation.

[1] For example,
http://crawl.akrasiac.org/rawdata/DreamWalker/crash-DreamWalker-20140611-033210.txt