New command: show weapons in viewport.
If a monster has a weapon, and you press ctrl-t (by default), you
will see, instead of its normal glyph, the glyph for the weapon it
is holding. Since this information is already available with 'x' and
even more easily in tiles, it's just an interface improvement, and as
far as I can tell in my tests there is no information leaked.