Don't allow selling items via Gozag wrath
Just give 1 gold per item destroyed, instead of tying it to the item's
value and allowing selling consumables (this could also be used for
price identification!).

The wrath effect is still very problematic in a lot of ways (allowing
dropping and picking up unwanted items repeatedly until they're affected
just to burn off penance, encouraging disabling autopickup especially for
scrolls like acquirement and enchant weapon that are particularly likely
to be affected and unlikely to be immediately needed).