Simplify spiny code, and ignore player armour.
The fact that lower-evp armour was better is nowhere obvious to the
player, and the spiny mutation is rare and low-impact enough that
most likely the player would never figure it out on their own. The
player now effectively wears ring mail all the time for trigger chance,
and a robe for damage.

There seemed to be an intention to make monster spines (only) ignore
half of AC, but since the argument was passed to the wrong place it
instead doubled the effective GDR; both are extremely marginal effects
and not worth keeping.

There was also a chance for monster spines to fail based on attacker
AC, in addition to checking AC normally for the damage; the chance
now depends only on the degree of monster spininess.