Add a prompt if a player tries to move with the Barbs status
Multiple people have died due to mindlessly moving and getting damaged.

Currently, you get the prompt once, and then a prop is set which
suppresses it in the future, so if you want to move a few squares to a
stairway you aren't spammed. This prop resets whenever the Barbs status
expires, so answering it affirmatively once won't affect another fight
later.