Properly handle map build failures in objstat
Previously an additional check on a level that builder() had accepted
could cause both mapstat and objstat to abort. We now only abort
mapstat in this case, because these failures aren't very relevant to
objstat. We also fail in objstat if builder() fails, since such levels
might be very problematic for object statistics, but continue on with
mapstat in this case.