Simplify and improve some initialization macros.
Make SET_FIELDS safer by using a do { } while(0) block, which also gives
us a scope, so we don't have to take the name of the intermediate array
variable as a parameter.  Rename it to INIT_VEC (it only works on
vectors of strings, but whatever).  While we're at it, also make sure
we evaluate the argument only once.  Also, use an array of const char *s
instead of strings, make that array static and const, and use the
ARRAYSZ macro.

Then, replace most of the uses of INIT_VEC née SET_FIELDS with a wrapper
macro ITEM_FILEDS that shortens things a bit.

Finally, move the macro definiton outside the function, and reformat.