Remove a puzzle vault
The diamond obelisk vault was added to demo Tornado, but in
practice it's a spoily puzzle vault. The diamond obelisk is
put to good use in a couple of decorative vaults, and at some
point might be interesting in a vault that uses it as terrain
instead of as a puzzle. This vault, though, has gotta go.