Decontaminate all chunks.
"This chunk gives you less nutrition" is not a very interesting effects,
and most of the time the only real effect is that you have to press 'e' a
few more times. If chunks on the whole are too common (which in the
current food environment is very hard to define) monsters in general can
give fewer chunks, or chunks can be less nutritious; no reason to make
some monsters special.

A good deal of code is still around because of the other less-nutrition
case, non-ghoul saprovores eating rotten chunks.