Remove sex demons
Retheme arenasprint's "succubus" to "dimme".

Not that it really matters, since no one actually plays
arenasprint (no one had seen this boss in more than a year),
but it was still kind of embarassing having it around.

The new sprite isn't ideal either, but I don't have the talent
to render "a hairy body, a lioness' head with donkey's teeth and
ears, long fingers and fingernails, and the feet of a bird with
sharp talons." Maybe someone else can!