Make octopode crushers a little more crushing
I've given them iron shot over icicle, and increased their HD by
two. This is more in keeping with the idea of them as the most
dangerous enemy in the Depths water population, roughly on par with
Tengu reavers as a threat. They've been slightly rethemed in the
description so as to mention the affinity with earth magic, which is
perhaps more appropriate for something that "crushes".