Have monsters check constriction before using knockback abilities
If a monster is constricting a victim, don't allow it to use a spell
or ability that would knock the victim back, thus breaking the
constriction.

I made this commit when I was considering giving op crushers primal
wave, an idea that I dropped, but the check should still be there, and
this does move otherwise duplicated code to a single method.