Merge pickup_menu and pickup_menu_limit.
Since pickup_menu = true was basically identical to pickup_menu_limit =
1 (or 2, actually), there isn't much need for two separate options. The
new default for pickup_menu_limit is 1, so there is no change in the
defaults.

However, if you pickup_menu_limit is now exclusive, not inclusive. This
fixes the weirdness that 1 and 2 meant the same thing, since you can't get
a menu with only one item on it. So, if you had pickup_menu = false, there
probably is a change in behaviour.