Improve behaviour on failure to monster teleport failure (#3839).
It's fairly easy to reproduce by wizmoding Zig:1 full of plants and
then teleporting a monster in there: it will probably end up in a wall.
As the mantis issue says, this will basically never happen in a real game,
but if it were to do so, just keeping it in place (with an appropriate
message seems like the best behaviour).