Add a lemuel_tele_altar variant.
This one will always let you in: you just have to find the correct side.
Split the weight with lemuel_tele_altar.

One problem is that monsters won't step on the teleporters, so it might
be too easy to get free kills (same as lava, except it works for flyers
and lava-swimmers too).