Don't disappear companions from Pan on load (크롤러, PleasingFungus)
This arose because of the combination of 0.12-a0-2710-g8500c20 with the
unmarshalling check originally added in 0.12-a0-2489-g1bd74f5.  We were
checking whether the companion was on-level, but in the case of Pan that
checked the mid cache (since all Pan levels have the same level_id).

I am not sure why the check is necessary at all: surely if the monster
is actually on another level, it shouldn't be alive in this level's menv
array?