Make a rotten animate dead message slightly less bad
The message for 'animate dead' turning rotten corpses
(necrophages/ghouls) into skeletons would appear when
monsters cast the spell, even if the corpse wasn't in
LOS. It's only intended to explain to the *player* why
they're not getting a zombie, so it's probably simpler
to just not print it for monster animate dead at all.

It will still print once for every applicable corpse,
which isn't ideal. Not sure how to fix it, though -
the relevant code is also, appropriately, pretty rotten...