Fix item_pickup triggers and rune/orb milestones.
The former were broken by 0.15-a0-1884-gb98b893 and the latter by
0.15-a0-1887-gff848d3.

Unfortunately, moving the floor quantity decrement later as this
commit does means there is more opportunity for the player to send
a HUP and duplicate items.  That was already a problem before
0.15-a0-1884-gb98b893, though, and I'd rather have the potential
exploit than have non-functional item_pickup triggers.

Milestones could be fixed without reintroducing the exploit, by using a
temporary on-stack copy of the object.  That doesn't work for the pickup
trigger, though, because the latter needs the object's index and
therefore requires a (still-extant) floor item.