Refactor plus-squashing code
Move it into the actual item-generation code, which is probably
a more sensible place for it to be.

Probably the actual behaviour should be rethought at some point;
it'd be nice to support item shops selling items with plusses
again, and it might be interesting to play around with generating
more high-plus no-ego items, which are very rare at the moment.
Might also make sense to tweak the threshold for aux armour up to
+2, to avoid people getting too jaded by glowing gloves/etc.