Make potions of cancellation a little more consistent
There was a very old special case preventing !cancel from
removing Slow caused by berserk. This was not only unintuitive
to players, but also inconsistent - the reasoning was that we
shouldn't affect nonmagical slowing, but what about slowing
from curare, which the potion does affect? What about slowing
which just happens to coincide with exhaust from another source?

Simpler to make it operate on a per-status-effect basis, rather
than trying to divine the original source of effects from
inadquate information.