Make blocked beams fall back to visible paths if possible (#8804).
In this position, you would be able to see the kobold, but trying to
shoot it would say your line of fire is blocked by the wall:

  ..#@
  K.8.

If the statue weren't there the beam would avoid the wall perfectly
well.  Furthermore, the "Aim" prompt does in fact say "fire blocked by a
granite statue", not "wall"; only the tracer and beam reported the
opaque wall as the reason for losing LOF.

When no LOS_SOLID_SEE ray is found for a beam, now we first try falling
back to a LOS_DEFAULT ray; only if that too fails do we take the
straight-path fallback.