Adjust volcano_overflow
It's unclear that players need to rest for the lava to recede, so always
leave a path around the outside leading back to the exit, and also always
make all of the created lava eventually recede to make sure it's not
possible to trap players. Make the lava rise a little more slowly,
slightly reduce the amount of loot, and add a few (flying) monsters.