Turn rings of regen into amulets
Rings of regen are notorious for encouraging tedious swapping,
since they're at best marginally useful in combat, and quite
useful outside combat. The amulet slot has a number of types
that already discourage swapping (faith, guardian spirit,
gourmand), which should at least ameliorate the regen problem
somewhat; and if we later want to change regen to do something
similar, it'll be a bit more intuitive if it's in the same slot.

This also helps refresh the pool of amulets, which was starting
to look a little shabby.