Allow giving derived undead equipment in mspecs (#8807)
Since they can be animated with equipment.  This doesn't check that the
item is something a zombie/skeleton can use, so it's up to vault
designers not to make an centaur zombie ; shortbow .

Items are still disallowed if the zombie's base monster can't use
equipment, of if its base type is random.