Remove Haste from spriggan air mages.
The justification given for giving them the spell initially in 0b45604
was "they're delayed in hasting themselves, causing them to be similar
in threat to their previous selves". It's been since established that
their very first action is often to haste themselves, leading to
Airstrike spam and causing them to be overwhelmingly more lethal than
anything else in Depths to the point where it's undesirable.