Don't veto when failing to connect Abyss vaults
So they'll actually place.

We should look more closely into *why* it fails to connect and whether
that can/should be fixed; or, alternatively, whether we should make
abyss vaults all no_exits.  For now, this stopgap at least leaves us in
the same situation we were in before the #8833 crashes began.