Forbid butchering/eating Xtahua and Gastronok under Zin (#7726)
Now, if the corpse or chunk has an orig_monnum, we use that rather
than the mon_type (which is just the species) for determining the
intelligence of the monster from which it came.

Mixing chunks from Xtahua and a normal fire dragon could result in a
stack that is either all edible, or all inedible (depending on which
stack was the source of the merge and which the target).  Furthermore,
these chunks are only marked by the forbidden tag, and only if you
worship Zin.  This could be improved, but it is not clear how worthwhile
that would be.  On the other hand, preserving names on chunks might be
"fun" for some players (even if that is kind of creepy).