add map view command to find the next explore target
This is useful for when you get the "Partly explored, can't reach some
places." message, but have no idea where the part that's unexplored
actually is (because it's in a corner behind a plant or something). It's
also useful if you don't actually want to use autoexplore (because there
are dangerous monsters out of line of sight, for instance), but you do
want to know which part of the level you still need to (carefully)
explore.