Make Gozag trunk-only
There's a general agreement that Gozag isn't ready for 0.15;
he needs a significant amount of design work. For now he will
languish in trunk, like the lava orcs of yore. (Unlike them, he
will almost certainly be in 0.16.)

This commit is extremely ugly & hacky & should be reverted as
soon as we branch.