List all food types on \.
Now that there are only 7 (including chunks), there's no reason to not
have as many as possible of a single type (because of weight being gone),
you often might want to just have one slot of non-chunk food. The
categories before weren't quite obvious (preserved meat didn't include
beef jerkies apparently). Now you can just set whichever food type you
found the most of to be picked up and everything else off.

The downside is that this clutters the menu. If it's too much, possibly
the categories could become chunks, meat, bread, fruit, and other,
splitting bread and fruit for a Ko of Fedhas, for example. That only saves
two lines on this, though.