Don't give out royal jellies in Spider (ends).
It's not necessary to have a reward in the branch end (Swamp has none, and
Snake has only the gear of the nagas), royal jellies aren't very connected
to Spider, and food isn't very good loot anyway.

The main reason the vaults got ambrosia replaced with royal jelly was
pending a review, so here one is.