Make potions of resistance grant rCorr
rCorr is less common with the removal of the preservation ego and
recently less effective at blocking the corrosion effect. Adding rCorr
to the potion increases the resistance's availability and falls
reasonably under the umbrella of "temporary natural resistances" of
the consumable. After the 0.15 release, we can consider if we want to
add rCorr as to artefacts or as an ego for e.g. cloaks, possibly
contingent on adding further monster acid attacks, in which case we
can reconsider the potion if need be.