Adjust ranged weapons slightly
Buff the hand crossbow & shortbow slightly to account for the
lack of ammo in the early-game (when they'll be used) - +1 base
dam apiece, and +1 base acc for the shortbow to make it 'feel' a
little more satisfying.

Adjust the triple crossbow very slightly downward - turns out the
'armour piercing' effect was more strong than expected, especially
when combining weapon & ammo brands. -1 base dam.