Make magic mapping reliably reveal the entire map
The 'patchy holes' in the map generally didn't hide anything
important, especially since in the most common cases when they
were relevant (portal vault entrances), you could generally look
at the surrounding tiles to guess where the entrance was. The
primary effect, in practice, was occasionally fucking with
autotravel.

Making *non*-patchy holes in the map would be an interesting
variant, but harder to implement; this is a simple improvement
to make in the meanwhile.