timeout tests on travis in 9 minutes, so we get a crash report
travis has a hard limit of 10 minutes with no output, so kill the tests
ourselves before it gets to that point, so that we can also get
morgues/crashdumps. could also solve this by having the loop output
something, but we really don't want to be running tests for that long
anyway.