Fix a crash with monsters drinking stacks of blood potions (qw).
The oldest blood timer needs to be removed before the quantity is changed,
since the timers might need be initialized before the oldest can be removed
and the initialization looks at the quantity.

I feel that ideally blood would always have initialized timers to avoid
issues like this, but the current blood code in rot.cc is clearly trying not
to assume this.