Replace an armour swap delay hack with an armour swap delay hack (qw).
The previous hack included 0.0-delay turns, which could lead to strange
behavior by calling various things in player_reacts() with no delay.
For instance, qw figured out how to abuse this to increment the turn
counter arbitrarily in 0 aut.

This hack also gets the turncount correct for armour swap actions:
now it takes 5 turns and 5.0 delay to take off a piece of armour
instead of 6 turns and 5.0 delay.