Tutorial 3: Adjust for jewellery id-on-wear.
Instead of using a scroll of identify to recognize the amulet's uselessness, the amulet (of inaccuracy) is now cursed, and is uncursed with a scroll of remove curse.

This was the path of least resistance, which keeps amulet wearing and item dropping. There's now no identification in the tutorial. Perhaps there could be some potion identification.