remove the high score cap (8884)
It has never been hit in a normal game (although the current online high
score is 65369553, which is close), but it's hit occasionally in sprint,
but inconsistently - it's only applied to games that don't provide their
own scoring system (so we have scores like 2147400000 in meatsprint, for
instance). I don't really see a purpose for it to begin with, and it's
been around since before stone soup (when the scoring algorithm was
entirely different), so I don't think it's particularly worth keeping.