Make the captain's cutlass disarmingly good.
It now disarms a monster's weapon with a dam / 75 chance; it shouldn't be
an effect that's happening every combat, just a cool bonus (like the
tendril disarming this is based on code- and idea-wise). The enchantment
is nerfed to +5, since it doesn't need it to be not strictly worse than
an ego item anymore.

Thanks to PleasingFungus for the description change!