Add god wrath descriptions
In an attempt to despoilerify mechanics that don't really need
to be spoilers, divine dislikes and wrath effects are now in a
new screen. (Except for Xom, who is Very Special.) I've made an
effort to clarify which gods care about abandonment, which gods
are 'evil' and 'chaotic', and which god wrath effects scale with
XL. (Mostly the monster summoning ones.)

It might make sense to expand some of the detailed god
descriptions (e.g. the description of Zin's recite), now that
the second screen is somewhat less crowded.