Remove dat/clua/wield.lua.
It duplicated some code that was already duplicated in C++ (also fixed in
this commit). I can't imagine much use for something that would display
items of your choice on the 'w' menu -- if something is useful, like
showing arrows/javs with s2s known, it should be on for all players.