A server game scores viewer for WebTiles
This is a client-side server score viewer that gets entries parsed from
the WebTiles server and send over JSON. Doing the rendering client-side
(as opposed to generating the html on the server) allows much better
consistency and integration with the rest of the client. It's admittedly
somewhat ugly right now, but the visual can be improved.

Thanks to dplusdplus, the lld server admin, whose server-side
implementation inspired this, and whose score parsing code I adapted.

Some TODO items:

* Selector for the number of scores to display (currently we default to
  100, but the URL can be edited to get more).
* Pagination: allow browsing through the scores.
* Handle lobby_html events (the banner and footer), either by displaying
  on the score page or only sending these messages to the lobby
  component.
* Clean up the visuals, moving necessary style items to their own CSS
  class.
* Possibly move some or all score field formatting to the client.