Ditch the commented options in init.txt.
They were confusing for beginners, who often didn't understand they had to
remove the beginning '#', experts, who often were not sure whether the
line had the default or the non-default value, and devs, who often failed
to keep them up to date, or give the the correct value, or who did not
know whether an option should be in there, alike.

This has been suggested before and I know it's not uncontroversial, but
elliptic and dpeg at least seemed to support the idea. Even if this ends
up getting reverted, the default init.txt could probably stand to be
shortened by sticking to often-changed options.