Add a "Cloud" status light and an option to control it.
It can sometimes be a bit ambiguous whether you're still in a cloud, and
people often request a way to check without pressing 'x'. The light is
darkgrey if the cloud does no damage to you and will have no harmful
effects. The option enabling it defaults to false on tiles builds, where
it is much easier to see if you're in a cloud.