Unbreak save compat.
Some TAG_MINOR_ versions were left in from the iashol branch; they were
not supposed to go in trunk. This wouldn't have broken anything major,
except that they were in the middle of the enum, causing the code in
tags.cc to get fouled up.

I'm removing them from the middle of the enum, which means saves made
between the merge and this commit won't load afterwards, probably;
thankfully, the public servers have not rebuilt yet.