Scale Ziggurat monsters per floor very harsly to number of completed zigs
This is it, this is the top end of raw danger Crawl can provide (if
certainly not relative danger). There's not much left besides re-arranging
the player and monsters, adding (even) more depth-scaling to each floor,
or throwing out the randomness and just placing solely top threats.

At zero ziggurats left on the last floor, there should be no differences
from before. At six ziggurats completed, smaller early floors can be
completely full of monsters, and the last floors can easily reach the hard,
game-limited monsters per floor cap of 700. At thirteen ziggurats, the
depth-scaling hd-pool that ziggurat floors rely on will be almost maxed
out per floor, and the scaling is essentially done. This might put some
stress on the game itself and the servers, though it should take a while
before people make it that far.

Rewards will slightly, slowly, and randomly scale up to completed zigs
not that they really matter. Both forms of scaling only happens for completing
a Zig (so leaving a floor early will avoid the monster scaling), but I'm not
greatly concerned: after the first zig, further zigs are for bragging rights
in the first place. Those who want to continue to relish the raw top power a
character can have against the raw top threats the game can provide can make
use of said trick, but indefinite zig raiding is not much of a recognizable
accomplishment following that trick.

There was some dev discussion on removing the guarantee on the Depths ziggurat
and making the sole source of ziggurats past the first one (in Pan) be
an evokable available only on Zig:27. I'd support such further changes.