Put the Hall of Blades on Elf:2 (and another vault onto Elf:1)
The Hall of Blades and the Elven Halls both struggle under a homogenous
specific monster set, and have been shrunken over time to the point of
removing the Hall of Blades in 0.15. I think there's still some value
in Blade, having made it a decent gamble for possibly useful equipment,
and it targets both branches homogenity problem to mix them together.
Normal speed, no-resists conjurers and summoners (with a handful of
variants) can welcome high speed omni-resistant melee monsters, and the
flavour of the elves making constructs is certainly readily possible, too.

The original layout is used as about a third to a half of Elf:2 on one hard
side of the floor; I also made a relatively-similiar , half-weight alternative
layout that is more distinctly split but with a more open entry section.
There's a mix of lesser elves alongside a smaller amount of dancing weapons,
and the previous dual high-quality weapons are still there. It is a wild swing
in content compared to the rest of the branch or even the branch end vault,
so I've re-purposed old Blade branch entry vaults as a connecting part between
the two, with a runed door and some weak dancing daggers and falchions to serve
as warning transition (alongside whatever piles of blood, sword shapes, fancy
or straightforward pillars, and floor_vault floor tiles the these subvault may
contain). Random teleport may kinda skip this, and I have some light concerns
about communicating the reliable presence of this new Hall of Blades
incarnation, but players can adapt as they have adapted to bigger changes.
This is still a combination of two optional later branches, anyway, and there
is hardly much pressure involved as such.

Of course, between the Elf:3 endvault and this Elf:2 Blade vault, the first
floor looks a little sparse aside from the entry vault. There's also not much
room to fit Elf-specific vaults with those two floors occupied, either. As a
result, I'm making sure of one of a decent selection of Elf vaults is placed
on Elf:1, so that they're not completely lost in all these bigger vaults. The
selection is mostly based in placing visually distinct 098 vaults or vaults
playing up the higher elf classes, and is a little sparse for now even with
a few additions I threw together (at sixteen): more work can be done here to
make the numbers feel comfortable for their reliable placement, in a fashion
similiar to Swamp.

Sorry, infiniplex, for warping the new layouts so much.