Hellemental hellsprings (Bring back wellsprings solely within Cocytus, zigs)
Unique forms of debilitation are much maligned, and so are movement-binded
genuine dangers. Still, there are hypothetically contexts for it to work.
The hells, which encourage diving and rushing through its levels rather
than clearing them due to the hell effects, general low value of small
amounts of loot, and the high plateau of player power / xp in its position,
should be the most ideal of them. At the very least, said hell effects can
already dump weakling fodder onto the player and throw on a random negative
status, so a monster that can try to do a little better than that is hardly
out of place.

It also serves as part of (unique) non-demon non-undead "threats" for
extended, as is somewhat done currently with salamander firebrands in
Gehenna, iron golems in Dis, ghost crabs in Tartarus, and the mechanical
gap in resistances with Ds enemies versus demons for Pan. Cocytus is
lacking in its elemental and dangerous aspects to some extent when compared
to the other hells, anyway; some wellsprings sprinkled throughout the branch
and the end vaults are another equalizing buff step besides even more fiends.

(Also, coc_hangedman is slightly buffed due to its generous terrain compared
to the other Cocytus end vaults; the open water is closed in a little, and a
few more monsters are added in a few parts. The other ends were buffed slightly,
too, and all have some small use of wellsprings.)