Temporary quick-fix: Remove allies tracer on monster torment / tornado / chains
While 43baa798a0 does allow los-affecting spells to be cast out of the player's
sight for monsters, the tracer used is horrifically high and basically is
blocked by having about 6 hd of vulnerable allies near by even with an xl 27
player target. This mostly means next to no torment from deep elf demonologist
summons, when salamanders / giants / dragons are nearby in the Hells, when
demonspawn are nearby in Pan, and when a lich is encountered in Vaults / Depths
with friends, or most recently when bennu and emperor scorpions are nearby in
Tomb. Crawl's randomness, of course, made it seem like one was just getting
luck on no torment rolls.

As the quickest solution possible, I'm cutting the tracer for all monster
los checks's _foe_ratio, to replicate old behaviour but preserve use out of
player los. This incidentally also lets Lom tornado again without dragging
a wizard nearby, and the same with Nikola. It does mean they're ready to
hurt their allies, but this is a temporary in-term step. Ideally, much smarter
tracers are produced, with proximity for tornado and chain lightning, as well
as player current health for torment. Tracers could also ignore damaging
summons, since that's obviously also not much of an issue.